requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
fire==0.4.0
gunicorn==20.1.0
idna==3.4
jwt==1.3.1
libxml2-python==2.10.2
mock-open==1.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
urllib3==1.26.15
whitenoise==6.2.0
